% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_message_fur_color.R
\name{get_message_fur_color}
\alias{get_message_fur_color}
\title{Get a message with primary fur color}
\usage{
get_message_fur_color(primary_fur_color)
}
\arguments{
\item{primary_fur_color}{A character string including (oh surprise) a
primary fur color name.}
}
\value{
Nothing. It is used for its side effect of displaying a message.
}
\description{
Get a message with primary fur color
}
\examples{
get_message_fur_color(primary_fur_color = "Cinnamon")
}
